refactor: move consults related schemas into separate file
